+++ scripts/src/terraform/state.ts
@@ -101,13 +101,40 @@ export class State {
     this.setState(source)
   }
 
+  async reset() {
+    this.setState(await thisModule.loadState())
+  }
+
   async refresh() {
     if (env.TF_EXEC === 'true') {
       await cli.exec(`terraform refresh -lock=${env.TF_LOCK}`, undefined, {
         cwd: env.TF_WORKING_DIR
       })
     }
-    this.setState(await thisModule.loadState())
+    await this.reset()
+  }
+
+  getAllAddresses(): string[] {
+    const addresses = []
+    for (const resourceClass of ResourceConstructors) {
+      const classAddresses = this.getAddresses(resourceClass)
+      addresses.push(...classAddresses)
+    }
+    return addresses
+  }
+
+  getAddresses<T extends Resource>(
+    resourceClass: ResourceConstructor<T>
+  ): string[] {
+    if (ResourceConstructors.includes(resourceClass)) {
+      return (
+        this._state?.values?.root_module?.resources
+          .filter((r: any) => r.type === resourceClass.StateType)
+          .map((r: any) => r.address) || []
+      )
+    } else {
+      throw new Error(`${resourceClass.name} is not supported`)
+    }
   }
 
   getAllResources(): Resource[] {
@@ -134,10 +161,16 @@ export class State {
   }
 
   async addResource(id: Id, resource: Resource) {
+    await this.addResourceAt(id, resource.getStateAddress().toLowerCase())
+  }
+
+  async addResourceAt(id: Id, address: string) {
     if (env.TF_EXEC === 'true') {
-      const address = resource.getStateAddress().replaceAll('"', '\\"')
       await cli.exec(
-        `terraform import -lock=${env.TF_LOCK} "${address}" "${id}"`,
+        `terraform import -lock=${env.TF_LOCK} "${address.replaceAll(
+          '"',
+          '\\"'
+        )}" "${id}"`,
         undefined,
         {cwd: env.TF_WORKING_DIR}
       )
@@ -145,10 +178,16 @@ export class State {
   }
 
   async removeResource(resource: Resource) {
+    await this.removeResourceAt(resource.getStateAddress().toLowerCase())
+  }
+
+  async removeResourceAt(address: string) {
     if (env.TF_EXEC === 'true') {
-      const address = resource.getStateAddress().replaceAll('"', '\\"')
       await cli.exec(
-        `terraform state rm -lock=${env.TF_LOCK} "${address}"`,
+        `terraform state rm -lock=${env.TF_LOCK} "${address.replaceAll(
+          '"',
+          '\\"'
+        )}"`,
         undefined,
         {cwd: env.TF_WORKING_DIR}
       )
@@ -156,21 +195,19 @@ export class State {
   }
 
   async sync(resources: [Id, Resource][]) {
-    const oldResources = this.getAllResources()
-    for (const resource of oldResources) {
+    const addresses = this.getAllAddresses()
+    for (const address of addresses) {
       if (
         !resources.some(
-          ([_i, r]) => r.getStateAddress() === resource.getStateAddress()
+          ([_, r]) => r.getStateAddress().toLowerCase() === address
         )
       ) {
-        await this.removeResource(resource)
+        await this.removeResourceAt(address)
       }
     }
     for (const [id, resource] of resources) {
       if (
-        !oldResources.some(
-          r => r.getStateAddress() === resource.getStateAddress()
-        )
+        !addresses.some(a => a === resource.getStateAddress().toLowerCase())
       ) {
         await this.addResource(id, resource)
       }