+++ terraform/resources.tf
@@ -1,7 +1,7 @@
 resource "github_membership" "this" {
   for_each = merge([
     for role, members in lookup(local.config, "members", {}) : {
-      for member in members : "${member}" => {
+      for member in members : lower("${member}") => {
         username = member
         role     = role
       }
@@ -13,12 +13,13 @@ resource "github_membership" "this" {
 
   lifecycle {
     ignore_changes = []
+    prevent_destroy = true
   }
 }
 
 resource "github_repository" "this" {
   for_each = {
-    for repository, config in lookup(local.config, "repositories", {}) : repository => merge(config, {
+    for repository, config in lookup(local.config, "repositories", {}) : lower(repository) => merge(config, {
       name = repository
     })
   }
@@ -70,6 +71,7 @@ resource "github_repository" "this" {
 
   lifecycle {
     ignore_changes = []
+    prevent_destroy = true
   }
 }
 
@@ -78,7 +80,7 @@ resource "github_repository_collaborator" "this" {
     for repository, repository_config in lookup(local.config, "repositories", {}) :
     [
       for permission, members in lookup(repository_config, "collaborators", {}) : {
-        for member in members : "${repository}:${member}" => {
+        for member in members : lower("${repository}:${member}") => {
           repository = repository
           username   = member
           permission = permission
@@ -102,9 +104,9 @@ resource "github_branch_protection" "this" {
   for_each = merge([
     for repository, repository_config in lookup(local.config, "repositories", {}) :
     {
-      for pattern, config in lookup(repository_config, "branch_protection", {}) : "${repository}:${pattern}" => merge(config, {
+      for pattern, config in lookup(repository_config, "branch_protection", {}) : lower("${repository}:${pattern}") => merge(config, {
         pattern       = pattern
-        repository_id = github_repository.this[repository].node_id
+        repository_id = github_repository.this[lower(repository)].node_id
       })
     }
   ]...)
@@ -145,7 +147,7 @@ resource "github_branch_protection" "this" {
 
 resource "github_team" "this" {
   for_each = {
-    for team, config in lookup(local.config, "teams", {}) : team => merge(config, {
+    for team, config in lookup(local.config, "teams", {}) : lower(team) => merge(config, {
       name           = team
       parent_team_id = try(try(element(data.github_organization_teams.this[0].teams, index(data.github_organization_teams.this[0].teams.*.name, config.parent_team_id)).id, config.parent_team_id), null)
     })
@@ -166,9 +168,9 @@ resource "github_team_repository" "this" {
     for repository, repository_config in lookup(local.config, "repositories", {}) :
     [
       for permission, teams in lookup(repository_config, "teams", {}) : {
-        for team in teams : "${team}:${repository}" => {
+        for team in teams : lower("${team}:${repository}") => {
           repository = repository
-          team_id    = github_team.this[team].id
+          team_id    = github_team.this[lower(team)].id
           permission = permission
         }
       }
@@ -194,8 +196,8 @@ resource "github_team_membership" "this" {
     for team, team_config in lookup(local.config, "teams", {}) :
     [
       for role, members in lookup(team_config, "members", {}) : {
-        for member in members : "${team}:${member}" => {
-          team_id  = github_team.this[team].id
+        for member in members : lower("${team}:${member}") => {
+          team_id  = github_team.this[lower(team)].id
           username = member
           role     = role
         }
@@ -220,10 +222,10 @@ resource "github_repository_file" "this" {
         for file, config in lookup(repository_config, "files", {}) : merge(config, {
           repository = repository
           file       = file
-          branch     = github_repository.this[repository].default_branch
+          branch     = github_repository.this[lower(repository)].default_branch
           content    = try(file("${path.module}/../files/${config.content}"), config.content)
         }) if contains(keys(config), "content")
-      ] : "${config.repository}/${config.file}" => config
+      ] : lower("${config.repository}/${config.file}") => config
     }
   ]...)